% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_to_SE.R
\name{acc_df_to_SummarizedExperiment}
\alias{acc_df_to_SummarizedExperiment}
\title{Convert Actigraphy data.frame to a SummarizedExperiment}
\usage{
acc_df_to_SummarizedExperiment(
  x,
  path_column = "file",
  assay_name = path_column,
  template_images = NULL,
  ...
)
}
\arguments{
\item{x}{An optional `data.frame` describing the samples/participants.}

\item{path_column}{column in `x` that is the paths to the file
names of the data.  Can also be a list column of `x` where this column
are the images}

\item{assay_name}{Name of the `assay` in the
\code{\link{SummarizedExperiment}}}

\item{...}{additional arguments to pass to \code{\link{SummarizedExperiment}}}
}
\value{
A \code{SummarizedExperiment} output
}
\description{
Convert Actigraphy data.frame to a SummarizedExperiment
}
\examples{
file = system.file("extdata",
"TAS1H30182785 (2019-09-17).gt3x",
package = "SummarizedActigraphy")
files = rep(file, 10)
df = data.frame(file = files,
age = stats::rpois(length(files), 50))
se = acc_df_to_SummarizedExperiment(df, "file")
}
