% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actigraphy_to_SE.R
\name{actigraphy_df_to_SummarizedExperiment}
\alias{actigraphy_df_to_SummarizedExperiment}
\title{Convert Actigraphy data.frame to a SummarizedExperiment}
\usage{
actigraphy_df_to_SummarizedExperiment(
  x,
  path_column = "file",
  measure = NULL,
  assay_name = measure,
  rowData = NULL,
  ...,
  verbose = TRUE,
  metadata = list()
)
}
\arguments{
\item{x}{An optional `data.frame` describing the samples/participants.}

\item{path_column}{column in `x` that is the paths to the file
names of the data.  Can also be a list column of `x` where this column
are the images}

\item{measure}{measure to grab to create an assay.  If \code{NULL},
then every measure that is not the column \code{time}}

\item{assay_name}{Name of the `assay` in the
\code{\link{SummarizedExperiment}}}

\item{rowData}{row data to pass to \code{\link{SummarizedExperiment}}}

\item{...}{additional arguments to pass to \code{\link{summarize_actigraphy}}}

\item{verbose}{print diagnostic messages}

\item{metadata}{metadata passed to \code{\link{SummarizedExperiment}}}
}
\value{
A \code{SummarizedExperiment} output
}
\description{
Convert Actigraphy data.frame to a SummarizedExperiment
}
\examples{
file = system.file("extdata",
"TAS1H30182785_2019-09-17.gt3x",
package = "SummarizedActigraphy")
files = rep(file, 2)
df = data.frame(file = files,
age = stats::rpois(length(files), 50),
stringsAsFactors = FALSE)
se = actigraphy_df_to_SummarizedExperiment(df, "file")
\dontrun{
df$file = factor(df$file)
se = actigraphy_df_to_SummarizedExperiment(df, "file", measure = "AI_mean")
}

}
