% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_measures.R
\name{calculate_measures}
\alias{calculate_measures}
\alias{calculate_ai}
\alias{calculate_n_idle}
\alias{calculate_mad}
\alias{calculate_mims}
\title{Calculate Summary Measures from Raw Accelerometer Data}
\usage{
calculate_measures(
  df,
  unit = "1 min",
  fix_zeros = TRUE,
  fill_in = TRUE,
  trim = FALSE,
  dynamic_range = c(-6, 6),
  calculate_mims = TRUE,
  verbose = TRUE,
  ...
)

calculate_ai(df, unit = "1 min")

calculate_n_idle(df, unit = "1 min")

calculate_mad(df, unit = "1 min")

calculate_mims(df, unit = "1 min", dynamic_range = c(-6, 6), ...)
}
\arguments{
\item{df}{An object with columns `X`, `Y`, and `Z` or an
object of class `AccData`}

\item{unit}{length of time to calculate measures over.  a character string
specifying a time unit or a multiple of a unit to be rounded to.
Valid base units are `second`, `minute`, `hour`, `day`, `week`, `month`,
`bimonth`, `quarter`, `season`, `halfyear`, and `year`.
Arbitrary unique English abbreviations as in the \code{\link{period}}
constructor are allowed.}

\item{fix_zeros}{Should \code{\link{fix_zeros}} be run before calculating
the measures?}

\item{fill_in}{if \code{fix_zeros = TRUE}, should the zeros be
filled in with the last}

\item{trim}{if \code{fix_zeros = TRUE},
should the time course be trimmed for zero values at
the beginning and the end of the time course?
observation carried forward?}

\item{dynamic_range}{Dynamic range of the device, in gravity units}

\item{calculate_mims}{Should MIMS units be calculated?}

\item{verbose}{print diagnostic messages}

\item{...}{additional arguments to pass to [MIMSunit::mims_unit]}
}
\value{
A data set with the calculated features
}
\description{
Calculate Summary Measures from Raw Accelerometer Data
}
