% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_measures.R
\name{ensure_header_timestamp}
\alias{ensure_header_timestamp}
\title{Ensure `HEADER_TIME_STAMP` is in}
\usage{
ensure_header_timestamp(df, subset = TRUE)
}
\arguments{
\item{df}{A data set of time and X/Y/Z}

\item{subset}{select only the columns of time and X/Y/Z}
}
\value{
A data set/`tibble`.  It does not return an `AccData` object
}
\description{
Ensure `HEADER_TIME_STAMP` is in
}
\examples{
file = system.file("extdata",
"TAS1H30182785_2019-09-17.gt3x",
package = "SummarizedActigraphy")
res = read_actigraphy(file, verbose = FALSE)
df = ensure_header_timestamp(res)
}
