% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_spike.R
\name{flag_qc}
\alias{flag_qc}
\alias{flag_qc_all}
\title{Flag Quality Control Values}
\usage{
flag_qc(
  df,
  dynamic_range = NULL,
  verbose = TRUE,
  flags = c("all", "spike", "interval_jump", "spike_second", "same_value",
    "device_limit", "all_zero", "impossible")
)

flag_qc_all(
  df,
  dynamic_range = NULL,
  verbose = TRUE,
  flags = c("all", "spike", "interval_jump", "spike_second", "same_value",
    "device_limit", "all_zero", "impossible")
)
}
\arguments{
\item{df}{A data set of actigraphy}

\item{dynamic_range}{dynamic range of the device, used to find the
device limit.}

\item{verbose}{print diagnostic messages}

\item{flags}{the flags to run for QC.  If you set this to \code{"all"},
then all flags are run, as the default.}
}
\value{
A data set with a `flags` column (`flag_qc`) or a number of
columns starting with `flag_*` (`flag_qc_all`)
}
\description{
Flag Quality Control Values
}
\examples{
file = system.file("extdata", "TAS1H30182785_2019-09-17.gt3x",
package = "SummarizedActigraphy")
res = read_actigraphy(file)
out = flag_qc(res)
}
