% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_measures.R
\name{get_dynamic_range}
\alias{get_dynamic_range}
\title{Get Dynamic Range}
\usage{
get_dynamic_range(df, dynamic_range = NULL)
}
\arguments{
\item{df}{An \code{AccData} object from \code{\link{read_actigraphy}}}

\item{dynamic_range}{the dynamic range.  If this is not \code{NULL}, then
it will be guess from the header or the data}
}
\value{
A length-2 numeric vector, or the original dynamic range (no
checking done )
}
\description{
Get Dynamic Range
}
