% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark_condition.R
\name{mark_condition}
\alias{mark_condition}
\title{Mark a Condition of a Specified Minimum Length}
\usage{
mark_condition(x, min_length = 1)
}
\arguments{
\item{x}{A logical vector}

\item{min_length}{minimum length, contiguous \code{TRUE} values required}
}
\value{
A logical vector
}
\description{
Mark a Condition of a Specified Minimum Length
}
\examples{
x = c(FALSE, TRUE, TRUE, FALSE, FALSE, rep(TRUE, 10), FALSE, rep(TRUE, 20))
mark_condition(x)
mark_condition(x, 2)
mark_condition(x, 5)
mark_condition(x, 15)
}
