% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mims_helpers.R
\name{mims_default_extrapolation}
\alias{mims_default_extrapolation}
\alias{mims_default_interpolation}
\alias{mims_default_filtering}
\title{Default MIMS worker functions}
\usage{
mims_default_extrapolation(df, dynamic_range = NULL)

mims_default_interpolation(df)

mims_default_filtering(df)
}
\arguments{
\item{df}{data set of data, usually time and X/Y/Z.  Usually from
\code{\link{read_actigraphy}}}

\item{dynamic_range}{dynamic range of the data.  Will be passed to
\code{\link{get_dynamic_range}}}
}
\value{
A data set of data
}
\description{
Default MIMS worker functions
}
