% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nifti_eBayes.R
\name{nifti_eBayes}
\alias{nifti_eBayes}
\alias{eBayes_to_images}
\title{Run `eBayes` on NIfTI Images}
\usage{
nifti_eBayes(
  imgs,
  mask = NULL,
  verbose = TRUE,
  ...,
  proportion = 0.01,
  stdev.coef.lim = c(0.1, 4),
  trend = FALSE,
  robust = FALSE,
  winsor.tail.p = c(0.05, 0.1),
  coef = NULL,
  adjust.method = "BH"
)

eBayes_to_images(
  fit,
  mask,
  verbose = TRUE,
  proportion = 0.01,
  stdev.coef.lim = c(0.1, 4),
  trend = FALSE,
  robust = FALSE,
  winsor.tail.p = c(0.05, 0.1),
  coef = NULL,
  adjust.method = "BH"
)
}
\arguments{
\item{imgs}{List of images or character of filenames}

\item{mask}{Image to subset the data}

\item{verbose}{Print diagnostic messages}

\item{...}{additional arguments to pass to \code{\link{lmFit}}}

\item{proportion}{numeric value between 0 and 1, assumed proportion of
voxels which are differentially expressed, passed to \code{\link{eBayes}}}

\item{stdev.coef.lim}{numeric vector of length 2, assumed
lower and upper limits for the standard deviation of
log2-fold-changes for differentially expressed voxels,
passed to \code{\link{eBayes}}}

\item{trend}{ogical, should an intensity-trend be allowed for the
prior variance? Default is that the prior variance is constant,
passed to \code{\link{eBayes}}}

\item{robust}{logical, should the estimation of df.prior and
var.prior be robustified against outlier sample variances,
passed to \code{\link{eBayes}}}

\item{winsor.tail.p}{numeric vector of length 1 or 2,
giving left and right tail proportions of x to Winsorize.
Used only when \code{robust=TRUE}, passed to \code{\link{eBayes}}}

\item{coef}{column number or column name specifying which
coefficient or contrast of the linear model is of interest,
passed to \code{\link{topTable}}}

\item{adjust.method}{method to adjust p-value,
passed to \code{\link{topTable}}}

\item{fit}{an MArrayLM fitted model object produced by `lmFit` or
`contrasts.fit`.}
}
\value{


A list of output images from `eBayes` and the linear fit
}
\description{
Run `eBayes` on NIfTI Images
}
\examples{
tarfile = system.file("extdata", "can.tar.gz", package = "limmi")
tarfile
exdir = tempfile()
dir.create(exdir)
files = untar(tarfile = tarfile, list = TRUE, exdir = exdir)
files = files[!grepl("^[.]", basename(files))]
unz = untar(tarfile = tarfile, files = files, exdir = exdir)
files = files[ grepl("hdr", basename(files))]
files
files = file.path(exdir, files)
res = nifti_eBayes(files, verbose = FALSE)
\dontrun{
design_mat = cbind(x1 = rep(c(0, 1), each = 6),
x2 = rep(c(0, 1, 0, 1), each = 3))
res2 = nifti_eBayes(files, design = design_mat, coef = 2)
}
}
