% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nifti_images_to_matrix.R
\name{nifti_images_to_matrix}
\alias{nifti_images_to_matrix}
\title{Transform set of images to matrix}
\usage{
nifti_images_to_matrix(imgs, mask = NULL, verbose = TRUE)
}
\arguments{
\item{imgs}{Vector of files or list of images (niftiImage, array, or nifti)}

\item{mask}{Binary image to subset the voxels}

\item{verbose}{print diagnostic messages}
}
\value{
Matrix of V by n, where V is the product of the dimensions of one
image or the number of voxels in the mask, and n is the number of images
}
\description{
Creates a matrix, where the voxels are on the rows
and images are on the columns
}
