% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nifti_lmFit.R
\name{nifti_lmFit}
\alias{nifti_lmFit}
\title{Run `lmFit` on NIfTI Images}
\usage{
nifti_lmFit(imgs, mask = NULL, verbose = TRUE, ...)
}
\arguments{
\item{imgs}{List of images or character of filenames}

\item{mask}{Image to subset the data}

\item{verbose}{Print diagnostic messages}

\item{...}{additional arguments to pass to \code{\link{lmFit}}}
}
\value{
The result from \code{\link{lmFit}}
}
\description{
Run `lmFit` on NIfTI Images
}
