% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_accel_data.R
\name{resample_accel_data}
\alias{resample_accel_data}
\alias{resample_accel_data_to_time}
\title{Resample 3-axial input signal to a specific sampling rate}
\usage{
resample_accel_data(data, sample_rate, method = "linear", ...)

resample_accel_data_to_time(data, times, method = "linear", ...)
}
\arguments{
\item{data}{A `data.frame` with a column for time in `POSIXct` (usually
`time`), and `X`, `Y`, `Z`}

\item{sample_rate}{sampling frequency, coercible to an integer.
This is the sampling rate you're sampling the data *into*.}

\item{method}{method for interpolation. Options are
`"linear"/"constant"`, which uses `stats::approx`, or one of
`"fmm", "periodic", "natural", "monoH.FC", "hyman"`, which uses
`stats::spline`}

\item{...}{additional arguments to pass to [stats::approx()] or
[stats::spline]}

\item{times}{a vector of `POSIXct` date/time values to interpolate
the data to}
}
\value{
A `data.frame`/`tibble` of `time` and `X`, `Y`, `Z`.
}
\description{
Resample 3-axial input signal to a specific sampling rate
}
\examples{
options(digits.secs = 3)
gt3x_file = system.file("extdata/TAS1H30182785_2019-09-17.gt3x", package = "SummarizedActigraphy")
x = read_actigraphy(gt3x_file)$data
res = resample_accel_data(data = x, sample_rate = 80)
res = resample_accel_data(data = x, sample_rate = 100)
res = resample_accel_data(data = x, sample_rate = 1)
res = resample_accel_data_to_time(
  data = x,
  times = lubridate::floor_date(x$time, unit = "1 sec"),
)
res_nat = resample_accel_data_to_time(
  data = x,
  times = lubridate::floor_date(x$time, unit = "1 sec"),
  method = "natural"
)
}
