% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_data.R
\name{standardize_data}
\alias{standardize_data}
\alias{standardise_data}
\title{Standardize the Accelerometry Data}
\usage{
standardize_data(data, subset = TRUE)

standardise_data(data, subset = TRUE)
}
\arguments{
\item{data}{A `data.frame` with a column for time in `POSIXct` (usually
`time`), and `X`, `Y`, `Z`}

\item{subset}{should only the `HEADER_TIMESTAMP` (if available)
and `XYZ` be subset?}
}
\value{
A `data.frame` with `X/Y/Z` and a time in
`time` (if available).
}
\description{
Standardize the Accelerometry Data
}
