% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_day_data.R
\name{summarize_daily_actigraphy}
\alias{summarize_daily_actigraphy}
\alias{summarise_daily_actigraphy}
\alias{summarize_actigraphy}
\alias{summarise_actigraphy}
\title{Summarize Actigraphy Data}
\usage{
summarize_daily_actigraphy(
  x,
  units = "1 min",
  fix_zeros = TRUE,
  fill_in = TRUE,
  trim = FALSE,
  verbose = TRUE,
  ...
)

summarise_daily_actigraphy(
  x,
  units = "1 min",
  fix_zeros = TRUE,
  fill_in = TRUE,
  trim = FALSE,
  verbose = TRUE,
  ...
)

summarize_actigraphy(
  x,
  units = "1 min",
  .fns = list(mean = mean, median = median),
  verbose = TRUE,
  ...
)

summarise_actigraphy(
  x,
  units = "1 min",
  .fns = list(mean = mean, median = median),
  verbose = TRUE,
  ...
)
}
\arguments{
\item{x}{an AccData object.  If `x` is a character, then
\code{\link{read_actigraphy}} will be run}

\item{units}{units to group the data to take the statistic over}

\item{fix_zeros}{Should \code{\link{fix_zeros}} be run before calculating
the measures?}

\item{fill_in}{if \code{fix_zeros = TRUE}, should the zeros be
filled in with the last
observation carried forward?}

\item{trim}{if \code{fix_zeros = TRUE},
should the time course be trimmed for zero values at
the beginning and the end of the time course?
observation carried forward?}

\item{verbose}{print diagnostic messages}

\item{...}{Additional arguments to pass to
\code{\link{read_actigraphy}}}

\item{.fns}{Functions to apply to each of the selected columns.
See \code{\link{across}}}
}
\value{
A \code{tsibble} object, with 86400 rows,
with one row for each secon d of the day `24*60*60`.
}
\description{
Summarize Actigraphy Data
}
\examples{
url = "https://github.com/THLfi/read.gt3x/files/3522749/GT3X\%2B.01.day.gt3x.zip"
destfile = tempfile(fileext = ".zip")
dl = utils::download.file(url, destfile = destfile)
gt3x_file = utils::unzip(destfile, exdir = tempdir())
gt3x_file = gt3x_file[!grepl("__MACOSX", gt3x_file)]
path = gt3x_file

x = read_actigraphy(path)
dob = x$header$Value[x$header$Field == "DateOfBirth"]
if (length(dob) > 0) {
 SummarizedActigraphy:::ticks2datetime(dob)
}

library(dplyr)
library(lubridate)
options(digit.secs = 2)

fixed = fix_zeros(x)
\dontrun{
daily = summarize_daily_actigraphy(fixed, fix_zeros = FALSE)
}
average_day = summarize_actigraphy(fixed, fix_zeros = FALSE)
\dontrun{
if (requireNamespace("ggplot2", quietly = TRUE)) {
library(magrittr)
average_day \%>\%
  ggplot(aes(x = time, y = ai_mean)) +
  geom_line()

average_day \%>\%
  ggplot(aes(x = time, y = ai_median)) +
  geom_line()
}
}
}
