% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummarizedActigraphy-package.R
\docType{package}
\name{SummarizedActigraphy-package}
\alias{SummarizedActigraphy}
\alias{SummarizedActigraphy-package}
\title{SummarizedActigraphy: Coerce 'Actigraphy' to Summarized Experiments}
\description{
Provides functions for coercing 'actigraphy' data into
    'SummarizedExperiment' objects used for genomic analysis.  This
    conversion enables the use of functions that work on the 
    'SummarizedExperiment' class.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/muschellij2/SummarizedActigraphy}
  \item \url{http://johnmuschelli.com/SummarizedActigraphy/}
  \item Report bugs at \url{https://github.com/muschellij2/SummarizedActigraphy/issues}
}

}
\author{
\strong{Maintainer}: John Muschelli \email{muschellij2@gmail.com} (\href{https://orcid.org/0000-0001-6469-1750}{ORCID})

}
\keyword{internal}
