% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_axes.R
\name{tidy_axes}
\alias{tidy_axes}
\title{Tidy axes to a long format}
\usage{
tidy_axes(df, only_xyz = TRUE)
}
\arguments{
\item{df}{An object with columns a time column `X`, `Y`, and `Z` or an
object of class `AccData`}

\item{only_xyz}{should the data be subset for only X/Y/Z values (and time)?}
}
\value{
A long data set with `time`, `axis`, and `value`
}
\description{
Tidy axes to a long format
}
\examples{
file = system.file("extdata", "TAS1H30182785_2019-09-17.gt3x",
package = "SummarizedActigraphy")
res = read_actigraphy(file)
long = tidy_axes(res)
}
