% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_acc_csv.R
\name{write_acc_csv}
\alias{write_acc_csv}
\title{Write ActiGraph-like Accelerometer CSV}
\usage{
write_acc_csv(x, file = tempfile(fileext = ".csv.gz"))
}
\arguments{
\item{x}{either an \code{AccData} or a \code{data.frame} of time and values.}

\item{file}{CSV file to write out}
}
\value{
The character path of the output file.
}
\description{
Write ActiGraph-like Accelerometer CSV
}
\examples{
file = system.file("extdata", "TAS1H30182785_2019-09-17.csv.gz",
package = "SummarizedActigraphy")
if (file.exists(file)) {
   x = read_acc_csv(file)
   outfile = write_acc_csv(x)
   out2 = read_acc_csv(outfile)
   all.equal(x$data, out2$data)
}
}
