% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xyz.R
\docType{data}
\name{xyz}
\alias{xyz}
\alias{xyzt}
\alias{txyz}
\title{Vector of X, Y, Z, and maybe time}
\format{
A vector with 3 or 4 elements, which are:
\describe{
\item{X}{value to extract X column}
\item{Y}{value to extract Y column}
\item{Z}{value to extract Y column}
\item{time}{value to extract time column}
}

An object of class \code{character} of length 5.

An object of class \code{character} of length 4.
}
\usage{
xyz

xyzt

txyz
}
\description{
Vector of X, Y, Z, and maybe time
}
\keyword{datasets}
