% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{estimate_calibration_values}
\alias{estimate_calibration_values}
\alias{calibrate}
\title{Calibrate Accelerometer Data using GGIR}
\usage{
estimate_calibration_values(
  file,
  verbose = TRUE,
  fix_zeros = TRUE,
  fill_in = TRUE,
  by_second = FALSE,
  trim = FALSE,
  ...
)

calibrate(
  file,
  verbose = TRUE,
  fix_zeros = TRUE,
  fill_in = TRUE,
  by_second = FALSE,
  trim = FALSE,
  round_after_calibration = TRUE,
  ...
)
}
\arguments{
\item{file}{Either a GT3X file, `AccData` object, or `data.frame` with
`X/Y/Z` and `time`}

\item{verbose}{print diagnostic messages, higher number result in higher verbosity}

\item{fix_zeros}{Should \code{\link{fix_zeros}} be run before calculating
the measures?}

\item{fill_in}{if \code{fix_zeros = TRUE}, should the zeros be
filled in with the last}

\item{by_second}{Should the last observation carried forward be done
only within the same second?}

\item{trim}{if \code{fix_zeros = TRUE},
should the time course be trimmed for zero values at
the beginning and the end of the time course?
observation carried forward?}

\item{...}{Additional arguments to pass to \code{\link{g.calibrate}}}

\item{round_after_calibration}{Should the data be rounded after calibration?
Will round to 3 digits}
}
\value{
A set of calibration coefficients
}
\description{
Calibrate Accelerometer Data using GGIR
}
\examples{
\donttest{
index = 2
files = read.gt3x::gt3x_datapath(index)
if (all(is.na(files))) {
   files = file.path(tempdir(), ".read.gt3x-data",
   read.gt3x::gt3x_filename(index))
}
path = files[1]
res = read_actigraphy(path)
cab = calibrate(res)
}
}
