% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_zeros.R
\name{fix_zeros}
\alias{fix_zeros}
\alias{idle_na_locf}
\title{Fix Zeros from Idle Sleep Mode}
\usage{
fix_zeros(df, fill_in = TRUE)

idle_na_locf(df)
}
\arguments{
\item{df}{An object with columns `X`, `Y`, and `Z` or an
object of class `AccData`}

\item{fill_in}{Should the zeros be filled in with the last
observation carried forward?}
}
\value{
A data set with the zeros filled in
}
\description{
Fix Zeros from Idle Sleep Mode
}
\examples{
df = data.frame(
  X = c(0.3/sqrt(0.5), rep(0, 3)),
  Y = c(0.4/sqrt(0.5), rep(0, 3)),
  Z = c(0.5/sqrt(0.5), rep(0, 3)),
  stringsAsFactors = FALSE)
fix_zeros(df)
fix_zeros(df, fill_in = FALSE)
}
