% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{get_transformations}
\alias{get_transformations}
\title{Get Transformations}
\usage{
get_transformations(df)
}
\arguments{
\item{df}{data set of data, usually time and X/Y/Z.  Usually from
\code{\link{read_actigraphy}}}
}
\value{
\code{\link{set_transformations}} returns the data, with the
`transformations` attribute updated and \code{\link{set_transformations}}
returns the attribute `transformations`
}
\description{
Get Transformations
}
