% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggir_process.R
\name{ggir_process}
\alias{ggir_process}
\title{Run GGIR Processing}
\usage{
ggir_process(df, unit = "1 min", calibrate = TRUE, verbose = TRUE, ...)
}
\arguments{
\item{df}{An object with columns `X`, `Y`, and `Z` or an
object of class `AccData`}

\item{unit}{length of time to calculate measures over.  a character string
specifying a time unit or a multiple of a unit to be rounded to.
Valid base units are `second`, `minute`, `hour`, `day`, `week`, `month`,
`bimonth`, `quarter`, `season`, `halfyear`, and `year`.
Arbitrary unique English abbreviations as in the \code{\link{period}}
constructor are allowed.}

\item{calibrate}{Run calibration, passed to \code{do.calibrate} in \code{GGIR}}

\item{verbose}{print diagnostic messages}

\item{...}{additional arguments to pass to \code{\link{g.shell.GGIR}}}
}
\value{
A list with all the output elements
}
\description{
Run GGIR Processing
}
\examples{
\dontrun{
url = paste0("https://github.com/THLfi/read.gt3x/files/",
"3522749/GT3X\%2B.01.day.gt3x.zip")
destfile = tempfile(fileext = ".zip")
dl = utils::download.file(url, destfile = destfile, mode = "wb")
gt3x_file = utils::unzip(destfile, exdir = tempdir())
gt3x_file = gt3x_file[!grepl("__MACOSX", gt3x_file)]
path = gt3x_file
res = read_actigraphy(path)
res = fix_zeros(res)
out = ggir_process(res)
}
}
