% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_acc_csv.R
\name{read_acc_csv}
\alias{read_acc_csv}
\title{Read ActiGraph Accelerometer CSV}
\usage{
read_acc_csv(file, ...)
}
\arguments{
\item{file}{CSV file to read in}

\item{...}{additional arguments to pass to \code{\link{read_csv}}}
}
\value{
A list of the header and the data set
}
\description{
Read ActiGraph Accelerometer CSV
}
