% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_actigraphy.R
\name{read_actigraphy}
\alias{read_actigraphy}
\title{Read Actigraphy File}
\usage{
read_actigraphy(file, ...)
}
\arguments{
\item{file}{file to read}

\item{...}{additional arguments to pass to function}
}
\value{
An accelerometer object from the function used to
read the file.
}
\description{
Read Actigraphy File
}
\examples{
url = "https://github.com/THLfi/read.gt3x/files/3522749/GT3X\%2B.01.day.gt3x.zip"
destfile = tempfile(fileext = ".zip")
dl = utils::download.file(url, destfile = destfile)
gt3x_file = utils::unzip(destfile, exdir = tempdir())
gt3x_file = gt3x_file[!grepl("__MACOSX", gt3x_file)]
path = gt3x_file
res = read_actigraphy(path)
file = system.file("extdata",
"TAS1H30182785 (2019-09-17).gt3x",
package = "SummarizedActigraphy")
res = read_actigraphy(file)
file = system.file("extdata",
"MECSLEEP17_left wrist_012854_2013-12-09 11-37-24.bin.gz",
package = "SummarizedActigraphy")
res = read_actigraphy(file)

file = "blah.exe"
testthat::expect_error(read_actigraphy(file))
}
