% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_day_data.R
\name{summarize_daily_actigraphy}
\alias{summarize_daily_actigraphy}
\alias{summarise_daily_actigraphy}
\alias{summarize_actigraphy}
\alias{collapse_daily_actigraphy}
\alias{summarise_actigraphy}
\title{Summarize Actigraphy Data}
\usage{
summarize_daily_actigraphy(
  x,
  unit = "1 min",
  fix_zeros = TRUE,
  fill_in = TRUE,
  trim = FALSE,
  verbose = TRUE,
  calculate_mims = FALSE,
  calculate_ac = FALSE,
  flag_data = TRUE,
  ensure_all_time = TRUE,
  flags = NULL,
  ...
)

summarise_daily_actigraphy(
  x,
  unit = "1 min",
  fix_zeros = TRUE,
  fill_in = TRUE,
  trim = FALSE,
  verbose = TRUE,
  calculate_mims = FALSE,
  calculate_ac = FALSE,
  flag_data = TRUE,
  ensure_all_time = TRUE,
  flags = NULL,
  ...
)

summarize_actigraphy(
  x,
  unit = "1 min",
  .fns = list(mean = mean, median = median),
  verbose = TRUE,
  ...
)

collapse_daily_actigraphy(
  x,
  .fns = list(mean = mean, median = median),
  verbose = TRUE
)

summarise_actigraphy(
  x,
  unit = "1 min",
  .fns = list(mean = mean, median = median),
  verbose = TRUE,
  ...
)
}
\arguments{
\item{x}{an AccData object.  If `x` is a character, then
\code{\link{read_actigraphy}} will be run}

\item{unit}{units to group the data to take the statistic over}

\item{fix_zeros}{Should \code{\link{fix_zeros}} be run before calculating
the measures?}

\item{fill_in}{if \code{fix_zeros = TRUE}, should the zeros be
filled in with the last
observation carried forward?}

\item{trim}{if \code{fix_zeros = TRUE},
should the time course be trimmed for zero values at
the beginning and the end of the time course?
observation carried forward?}

\item{verbose}{print diagnostic messages}

\item{calculate_mims}{Should MIMS units be calculated?
Passed to \code{\link{calculate_measures}}}

\item{calculate_ac}{Should Activity Counts from the \code{activityCounts}
package be calculated?}

\item{flag_data}{Should [SummarizedActigraphy::flag_qc()] be run?
It will be executed after \code{fix_zeros} before any measure
calculation}

\item{ensure_all_time}{if \code{TRUE}, then all times from the first to
last times will be in the output, even if data during that time was not
in the input}

\item{flags}{the flags to calculate,
passed to [SummarizedActigraphy::flag_qc()]}

\item{...}{Additional arguments to pass to
\code{\link{read_actigraphy}}}

\item{.fns}{Functions to apply to each of the selected columns.
See \code{\link{across}}}
}
\value{
A \code{tsibble} object, with 86400 rows,
with one row for each secon d of the day `24*60*60`.
}
\description{
Summarize Actigraphy Data
}
\examples{
path = system.file("extdata",
"TAS1H30182785_2019-09-17.gt3x",
package = "SummarizedActigraphy")

x = read_actigraphy(path)

options(digit.secs = 2)

fixed = fix_zeros(x)
daily = summarize_daily_actigraphy(fixed, fix_zeros = FALSE)
average_day = summarize_actigraphy(fixed, fix_zeros = FALSE)
\dontrun{
if (requireNamespace("ggplot2", quietly = TRUE)) {
library(magrittr)
average_day \%>\%
  ggplot(aes(x = time, y = ai_mean)) +
  geom_line()

average_day \%>\%
  ggplot(aes(x = time, y = ai_median)) +
  geom_line()
}
}
}
